% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuny_db.R
\name{cuny_db}
\alias{cuny_db}
\title{Connect to the CUNY Sports Database}
\usage{
cuny_db()
}
\value{
A DBI database connection.
}
\description{
Opens a connection to the SQLite database included with CUNYsports.
}
