% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuny_pull_table.R
\name{cuny_pull_table}
\alias{cuny_pull_table}
\title{Pull a Table from the CUNY Sports Database}
\usage{
cuny_pull_table(x)
}
\arguments{
\item{x}{A character string specifying the name of the table to retrieve.}
}
\value{
A data frame containing the requested table.
}
\description{
Retrieves a complete table from the SQLite database included with
the CUNYsports package and returns it as an R data frame.
}
\details{
Use [cuny_table_names()] to view the tables currently available in
the CUNY Sports database.
}
\examples{
cuny_pull_table("basketball_all_seasons_team_final")

}
