% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuny_table_info.R
\name{cuny_table_info}
\alias{cuny_table_info}
\title{View Information About a CUNY Sports Table}
\usage{
cuny_table_info(x)
}
\arguments{
\item{x}{A character string specifying the name of a table in the
CUNY Sports database.}
}
\value{
A tibble containing metadata and descriptive information
  about the requested table.
}
\description{
Returns metadata and descriptive information for a table available
in the CUNY Sports database.
}
\details{
Table information is retrieved from the `table_catalog` dataset
included with the CUNYsports package. Use [cuny_table_names()] to
view the names of tables currently available in the database.
}
\examples{
cuny_table_info("basketball_all_seasons_team_final")

}
