% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CUNYsports-package.R
\docType{package}
\name{CUNYsports-package}
\alias{CUNYsports}
\alias{CUNYsports-package}
\title{CUNYsports: CUNY Athletics Data}
\description{
Tools for accessing and exploring collegiate athletics data from the
City University of New York (CUNY).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nyc-open-data-lab.github.io/CUNYsports/}
  \item \url{https://github.com/nyc-open-data-lab/CUNYsports}
  \item Report bugs at \url{https://github.com/nyc-open-data-lab/CUNYsports/issues}
}

}
\author{
\strong{Maintainer}: Christian Martinez \email{c.martinez0@outlook.com} (\href{https://orcid.org/0009-0005-6026-6454}{ORCID}) (GitHub: martinezc1)

Authors:
\itemize{
  \item Christian Martinez \email{c.martinez0@outlook.com} (\href{https://orcid.org/0009-0005-6026-6454}{ORCID}) (GitHub: martinezc1)
}

}
\keyword{internal}
